refactor: 코드리뷰 반영 (네이밍변경)
